test(workbench): move test pages into separate folder
