+++ apps/workbench-testing-app/src/app/test-pages/bulk-navigation-test-page/bulk-navigation-test-page.component.ts
@@ -10,7 +10,8 @@
 
 import {Component} from '@angular/core';
 import {WorkbenchRouter} from '@scion/workbench';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {FormBuilder, FormGroup, ReactiveFormsModule, Validators} from '@angular/forms';
+import {SciFormFieldModule} from '@scion/components.internal/form-field';
 
 const VIEW_COUNT = 'viewCount';
 const CSS_CLASS = 'cssClass';
@@ -19,6 +20,11 @@ const CSS_CLASS = 'cssClass';
   selector: 'app-bulk-navigation-test-page',
   templateUrl: './bulk-navigation-test-page.component.html',
   styleUrls: ['./bulk-navigation-test-page.component.scss'],
+  standalone: true,
+  imports: [
+    SciFormFieldModule,
+    ReactiveFormsModule,
+  ]
 })
 export class BulkNavigationTestPageComponent {
