+++ projects/scion/e2e-testing/src/workbench/bulk-navigation.e2e-spec.ts
@@ -9,7 +9,7 @@
  */
 
 import {test} from '../fixtures';
-import {BulkNavigationTestPagePO} from './page-object/bulk-navigation-test-page.po';
+import {BulkNavigationTestPagePO} from './page-object/test-pages/bulk-navigation-test-page.po';
 import {expect} from '@playwright/test';
 
 test.describe('Bulk Navigation', () => {
